# Re-implemented the ecvt function.