… src/classlibnative/bcltype/CMakeLists.txt
… src/classlibnative/bcltype/number.h
… src/pal/inc/pal.h
… src/pal/src/cruntime/misctls.cpp
… src/pal/tests/palsuite/c_runtime/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/_ecvt/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/_ecvt/test1/CMakeLists.txt
… src/pal/tests/palsuite/c_runtime/_ecvt/test1/test1.cpp
… src/pal/tests/palsuite/c_runtime/_ecvt/test1/testinfo.dat
… src/pal/tests/palsuite/paltestlist_to_be_reviewed.txt
… src/pal/tests/palsuite/palverify.dat