@safern 2017-07-19
> @safern could you please take a look at the CI falure in ComponentModel?

Will do later :)