@tannergooding 2017-07-19
I'm going to give the implementation a review pass sometime later this evening.