@tannergooding 2017-07-19
General feedback: The long names (such as `realMantissa`) make it very hard to validate that the implementation follows the algorithm described in the paper (where they use `r`, for example).

It would be much easier to follow/validate if you used the variable names declared in the paper and had a brief excerpt at the top defining each variable.