@tannergooding 2017-07-19
still reviewing, but need to take a break for a bit. Will pick back up after dinner.