@mazong1123 2017-07-20
@tarekgh it's worth mentioning that some of the ideas and code were modified from http://www.ryanjuckett.com/programming/printing-floating-point-numbers/ , especially for how to estimate the first digital exponent and doing a heuristic divide. If they are not good to be adopted to .NET core, I need to remove them.