@mazong1123 2017-07-20
@tannergooding thanks for your feedback. I'll go through them tonight (in China). Thanks!