@mazong1123 2017-07-20
> General feedback: The long names (such as realMantissa) make it very hard to validate that the implementation follows the algorithm described in the paper (where they use r, for example).

It would be much easier to follow/validate if you used the variable names declared in the paper and had a brief excerpt at the top defining each variable.

How about do it reversely? I keep current meaningful variable name, and add a brief excerpt to show the relationship between the meaningful variable name and the name in the paper? Then we would not force the developer to read the paper before read the code (if they do not want to go deeper).