@mazong1123 2017-07-26
On Linux, each test case on new implementation are much faster than or equal to the old one except `0.0`. New implementation is -40% slower than the old one. Below are the perf table:

Number | Avarage time before | Avarage time after | delta percentage
-- | -- | -- | --
104234.343 | 170 | 49 | 71%
double.MinValue/2 | 382 | 178 | 53%
PI | 67 | 45 | 33%
E | 67 | 44 | 34%
MaxValue | 365 | 193 | 47%
MinValue | 365 | 194 | 47%
NaN | 3 | 3 | 0%
-∞ | 3 | 3 | 0%
∞ | 3 | 3 | 0%
Max Subnormal | 126 | 52 | 58%
Min Subnormal | 135 | 62 | 54%
middle Subnormal | 126 | 52 | 58%
0.0 | 10 | 14 | -40%
-0.0 | 135 | 62 | 54%

I'm trying to introduce `wmemset` to improve the performance. However, it seems like `Tizen armel` tests failed by this change. I guess I need to work another way to improve the performance of converting `0.0` on Linux.